Move ClassNames & SpanNames to .util package
